🔵 renaming files accordingly and simplifying redundant types
