+++ packages/backend/src/domain/api.ts
@@ -2,6 +2,7 @@ import { Result } from 'true-myth'
 import { Etablissement, ContactId } from './types'
 
 type EtablissementResult = Result<Etablissement, Error>
+
 export type fetchEtablissement = (siret: string) => Promise<EtablissementResult>
 
 export type postNewContact = (
